+++ IF/Classes/scene/titan/Titan.cpp
@@ -55,7 +55,7 @@ struct sModelPram{
 bool Titan::initWithFile(int tid)
 {
     _tid = tid;
-    int nModelScale = atoi(CCCommonUtils::getPropById(CC_ITOA(_tid), "scale").c_str());
+    float nModelScale = atof(CCCommonUtils::getPropById(CC_ITOA(_tid), "scale").c_str());
     // 设置当前node的contentSize 用于点击事件
     setContentSize(Size(500,550));
     setAnchorPoint(Vec2(0.5,0));
@@ -582,11 +582,11 @@ void Titan::turnFront()
         //_model->setRotation3D(Vec3(0,180,0));
         //end a by ljf
         // 面朝正方向
-        //_model->setRotation3D(Vec3(32, 39, -24));
+        //_model->setRotation3D(Vec3(38, 39, -24));
         // 面朝左90度
-//        _model->setRotation3D(Vec3(32, -51, 24));
+//        _model->setRotation3D(Vec3(38, -51, 24));
         // 面朝左90度
-//        _model->setRotation3D(Vec3(32, -6, 24));
+//        _model->setRotation3D(Vec3(38, -6, 24));
     }
 }
